implementation insertOrExecute in calculator controller

+++ CalculatorApplication/src/it/unisa/diem/se/calculatorapplication/entity/ComplexNumber.java
@@ -24,7 +24,7 @@ public class ComplexNumber {
 
     private double real, imaginary;
     
-    public ComplexNumber (float real) {
+    public ComplexNumber (double real) {
         this.real = real;
         this.imaginary = 0.0;
     }
